+++ cmd/login/runner.go
@@ -13,6 +13,8 @@ import (
 	"github.com/spf13/cobra"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/client-go/tools/clientcmd"
+
+	"github.com/giantswarm/kubectl-gs/pkg/kubeconfig"
 )
 
 type runner struct {
@@ -28,13 +30,12 @@ type runner struct {
 }
 
 type LoginOptions struct {
-	isMCLogin         bool
-	isWCLogin         bool
-	selfContainedMC   bool
-	selfContainedWC   bool
-	switchToMCcontext bool
-	switchToWCcontext bool
-	originContext     string
+	isWCClientCert            bool
+	selfContained             bool
+	selfContainedClientCert   bool
+	switchToContext           bool
+	switchToClientCertContext bool
+	originContext             string
 }
 
 func (r *runner) Run(cmd *cobra.Command, args []string) error {
@@ -55,26 +56,57 @@ func (r *runner) Run(cmd *cobra.Command, args []string) error {
 }
 
 func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) error {
-	var err error
-
-	// Try to reuse the current context if no MC argument is given
-	if !r.loginOptions.isMCLogin {
-		err = r.tryToReuseExistingContext(ctx)
+	switch len(args) {
+	// No arguments given - we try to reuse the existing context.
+	case 0:
+		err := r.tryToReuseExistingContext(ctx)
 		if err != nil {
 			return microerror.Mask(err)
 		}
-	} else {
-		// MC argument can be a kubernetes context name,
-		// installation code name, or happa/k8s api URL.
+	// One argument - This can be an existing context, an installation name or a login URL
+	case 1:
 		installationIdentifier := strings.ToLower(args[0])
-		err = r.handleMCLogin(ctx, installationIdentifier)
+		foundContext, err := r.findContext(ctx, installationIdentifier)
+		if IsContextDoesNotExist(err) && !strings.HasSuffix(installationIdentifier, kubeconfig.ClientCertSuffix) {
+			clientCertContext := kubeconfig.GetClientCertContextName(installationIdentifier)
+			fmt.Fprint(r.stdout, color.YellowString("No context named %s was found: %s\nLooking for context %s.\n", installationIdentifier, err, clientCertContext))
+			foundContext, err = r.findContext(ctx, clientCertContext)
+		}
 		if err != nil {
 			return microerror.Mask(err)
 		}
+		if !foundContext {
+			var tokenOverride string
+			if c, ok := r.flag.config.(*genericclioptions.ConfigFlags); ok && c.BearerToken != nil && len(*c.BearerToken) > 0 {
+				tokenOverride = *c.BearerToken
+			}
+			err = r.loginWithURL(ctx, installationIdentifier, true, tokenOverride)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+		}
+	// Two arguments - This is interpreted as an installation and a workload cluster
+	case 2:
+		installationIdentifier := strings.ToLower(strings.Join(args, "-"))
+		foundContext, err := r.findContext(ctx, installationIdentifier)
+		if IsContextDoesNotExist(err) && !strings.HasSuffix(installationIdentifier, kubeconfig.ClientCertSuffix) {
+			clientCertContext := kubeconfig.GetClientCertContextName(installationIdentifier)
+			fmt.Fprint(r.stdout, color.YellowString("No context named %s was found: %s\nLooking for context %s.\n", installationIdentifier, err, clientCertContext))
+			foundContext, err = r.findContext(ctx, clientCertContext)
+		}
+		if err != nil {
+			return microerror.Mask(err)
+		}
+		if !foundContext {
+			return microerror.Maskf(contextDoesNotExistError, "Could not find context for identifier %s", installationIdentifier)
+		}
+	default:
+		return microerror.Maskf(invalidConfigError, "Invalid number of arguments.")
 	}
-	// Login to WC after MC if desired
-	if r.loginOptions.isWCLogin {
-		return r.handleWCLogin(ctx)
+
+	// Clientcert creation if desired
+	if r.loginOptions.isWCClientCert {
+		return r.handleWCClientCert(ctx)
 	}
 
 	return nil
@@ -106,14 +138,13 @@ func (r *runner) setLoginOptions(ctx context.Context, args *[]string) {
 		*args = append(*args, contextFlag)
 	}
 	r.loginOptions = LoginOptions{
-		originContext:   originContext,
-		isWCLogin:       len(r.flag.WCName) > 0,
-		isMCLogin:       !(len(*args) < 1),
-		selfContainedMC: len(r.flag.SelfContained) > 0 && !(len(r.flag.WCName) > 0),
-		selfContainedWC: len(r.flag.SelfContained) > 0 && len(r.flag.WCName) > 0,
+		originContext:           originContext,
+		isWCClientCert:          len(r.flag.WCName) > 0,
+		selfContained:           len(r.flag.SelfContained) > 0 && !(len(r.flag.WCName) > 0),
+		selfContainedClientCert: len(r.flag.SelfContained) > 0 && len(r.flag.WCName) > 0,
 	}
-	r.loginOptions.switchToMCcontext = r.loginOptions.isWCLogin || !(r.loginOptions.selfContainedMC || r.flag.KeepContext)
-	r.loginOptions.switchToWCcontext = r.loginOptions.isWCLogin && !(r.loginOptions.selfContainedWC || r.flag.KeepContext)
+	r.loginOptions.switchToContext = r.loginOptions.isWCClientCert || !(r.loginOptions.selfContained || r.flag.KeepContext)
+	r.loginOptions.switchToClientCertContext = r.loginOptions.isWCClientCert && !(r.loginOptions.selfContainedClientCert || r.flag.KeepContext)
 }
 
 func (r *runner) tryToReuseExistingContext(ctx context.Context) error {
@@ -125,6 +156,6 @@ func (r *runner) tryToReuseExistingContext(ctx context.Context) error {
 	if currentContext != "" {
 		return r.loginWithKubeContextName(ctx, currentContext)
 	} else {
-		return microerror.Maskf(selectedContextNonCompatibleError, "The current context does not seem to belong to a Giant Swarm management cluster.\nPlease run 'kubectl gs login --help' to find out how to log in to a particular management cluster.")
+		return microerror.Maskf(selectedContextNonCompatibleError, "The current context does not seem to belong to a Giant Swarm cluster.\nPlease run 'kubectl gs login --help' to find out how to log in to a particular cluster.")
 	}
 }